requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
Pillow>=10.0.0
easyocr>=1.7.0
